libxo: eliminate diff from apparent mis-merge

+++ contrib/libxo/libxo/xo_encoder.c
@@ -139,7 +139,7 @@ xo_encoder_list_add (const char *name)
 
     xo_encoder_node_t *xep = xo_realloc(NULL, sizeof(*xep));
     if (xep) {
-	int len = strlen(name) + 1;
+	ssize_t len = strlen(name) + 1;
 	xep->xe_name = xo_realloc(NULL, len);
 	if (xep->xe_name == NULL) {
 	    xo_free(xep);
